moved rotation out of Util class
